Worked unpack.go and clean dir.
